Update README.md
So With the text I just edited, I am presented two options. Commit directly to the main branch or create a new branch for this comment and start a pull request. I don't know what any of that means. I hope I am not interfering with the developer that I forked this from.